Revert "move receipt related helpers to another file"
This reverts commit c26f537cbb535d4a3f698d1a6a606555aad85cef.
