+++ packages/transactions/src/index.ts
@@ -1,6 +1,7 @@
 export * from './coders/input';
 export * from './coders/output';
 export * from './coders/policy';
+export * from './coders/receipt';
 export * from './coders/storage-slot';
 export * from './coders/transaction';
 export * from './coders/witness';
@@ -8,4 +9,3 @@ export * from './coders/tx-pointer';
 export * from './coders/utxo-id';
 export * from './coders/byte-array';
 export * from './receipts';
-export * from './utils';
